requirements.txt
aenum==2.2.6
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
paho-mqtt==1.6.1
pyotp==2.6.0
python-dotenv==0.20.0
requests==2.28.0
soupsieve==2.3.2.post1
urllib3==1.26.9
